Added maple samplers documentation
The maple files distance_hereditary_sampler.mpl and
three_leaf_power_sampler.mpl contain implementations of unbiased
samplers for, respectively, the classes of (split-trees of) unlabeled
distance-hereditary and three-leaf power graphs. sampler_tools.mpl
contains helper functions for these two sampler files, as well as tools
to estimate the radius of convergence of a power series and count the
number of nodes of a certain type in a split tree.